+++ src/main/kotlin/model/Objective.kt
@@ -4,7 +4,7 @@ import model.auxiliary.ObjectiveType
 
 // model.Objective
 
-abstract class Objective (val id: String, val skill: Skill, val minutes: Int)  {
+abstract class Objective (val id: String, val skill: Skill, val minutes: Double)  {
     abstract val objectiveType: ObjectiveType
     abstract fun printShort(startLevel: Int)  // print a one line text representation of the objective
     abstract fun log(value: Double)
